pyproject.toml
[tool.poetry]
name = "layer-architecture"
version = "0.1.0"
description = ""
authors = ["LJ Li <ljli@xrex.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.93.0"
uvicorn = "^0.20.0"
alembic = "^1.10.2"
asgi-correlation-id = "^4.1.0"
python-dotenv = "^1.0.0"
mysqlclient = "^2.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pre-commit = "^3.1.1"
httpx = "^0.23.3"
mypy = "^1.1.1"
flake8 = "^6.0.0"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
include = ["./app"]

[tool.pytest.ini_options]
pythonpath = [
    ".", "app"
]

[tool.bandit]
exclude_dirs = ["tests"]

